+++ pydaynight/__init__.py
@@ -1,4 +1,4 @@
 from . import _version
-from .pydaynight import julian_day
+from .pydaynight import julian_day, sun_angle
 
 __version__ = _version.get_versions()["version"]